pyproject.toml
[tool.poetry]
name = "autofe"
version = "0.1.0"
description = ""
authors = ["eppou <vinipre7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
xarray = "^2025.12.0"
netcdf4 = "^1.7.4"
numpy = "^2.4.1"
pandas = "^2.3.3"
pyyaml = "^6.0.3"
dask = "^2026.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
